# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_post",
        "type": "object"
      },
...
      "DocumentInfo": {
        "additionalProperties": false,
        "properties": {
          "added_at": {
            "format": "date-time",
            "title": "Added At",
            "type": "string"
          },
          "chunks": {
            "title": "Chunks",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "size_bytes",
          "chunks",
          "added_at"
        ],
        "title": "DocumentInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeetingListItem": {
        "additionalProperties": false,
        "description": "Field order is the sessions-list display order: follow-ups before alerts.",
        "properties": {
          "alert_count": {
            "title": "Alert Count",
            "type": "integer"
          },
          "follow_ups_outstanding": {
            "title": "Follow Ups Outstanding",
            "type": "integer"
          },
          "follow_ups_resolved": {
            "title": "Follow Ups Resolved",
            "type": "integer"
          },
          "meeting_id": {
            "title": "Meeting Id",
            "type": "string"
          },
          "participants": {
            "items": {
              "type": "string"
            },
            "title": "Participants",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "meeting_id",
          "title",
          "started_at",
          "participants",
          "follow_ups_outstanding",
          "follow_ups_resolved",
          "alert_count"
        ],
        "title": "MeetingListItem",
        "type": "object"
      },
...
      "Ok": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "Ok",
        "type": "object"
      },
...
      "ReplayMeetingRequest": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "default": "Fake meeting (replay)",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "ReplayMeetingRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Google Meet copilot backend. Contract: DESIGN.md \u00a74.",
    "title": "Meet AGI",
    "version": "0.1.0-milestone-0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dev/meetings": {
      "post": {
        "operationId": "create_replay_meeting_api_dev_meetings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayMeetingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Replay Meeting"
      }
    },
    "/api/documents": {
      "get": {
        "operationId": "list_documents_api_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentInfo"
                  },
                  "title": "Response List Documents Api Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Documents"
      },
      "post": {
        "operationId": "upload_document_api_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document"
      }
    },
    "/api/documents/{name}": {
      "delete": {
        "operationId": "delete_document_api_documents__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Document"
      }
    },
...
    "/api/meetings": {
      "get": {
        "operationId": "list_meetings_api_meetings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeetingListItem"
                  },
                  "title": "Response List Meetings Api Meetings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Meetings"
      },
      "post": {
        "operationId": "create_meeting_api_meetings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeetingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Meeting"
      }
    },
    "/api/meetings/{meeting_id}": {
      "get": {
        "operationId": "get_meeting_api_meetings__meeting_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Meeting"
      }
    },
    "/api/meetings/{meeting_id}/end": {
      "post": {
        "operationId": "end_meeting_api_meetings__meeting_id__end_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End Meeting"
      }
    },
    "/api/meetings/{meeting_id}/events": {
      "get": {
        "operationId": "events_api_meetings__meeting_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Since",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Server-Sent Events; each `data:` line is one Event (schema: GET /api/contract/events)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events"
      }
    },
    "/api/meetings/{meeting_id}/follow-ups/{follow_up_id}": {
      "patch": {
        "operationId": "patch_follow_up_api_meetings__meeting_id__follow_ups__follow_up_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "follow_up_id",
            "required": true,
            "schema": {
              "title": "Follow Up Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUp"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Follow Up"
      }
    },
    "/api/meetings/{meeting_id}/mute": {
      "post": {
        "operationId": "mute_api_meetings__meeting_id__mute_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "alert": "#/components/schemas/AlertEvent",
                      "bot.status": "#/components/schemas/BotStatusEvent",
                      "chat.post": "#/components/schemas/ChatPostEvent",
                      "follow_up": "#/components/schemas/FollowUpEvent",
                      "meeting.ended": "#/components/schemas/MeetingEndedEvent",
                      "meeting.summary": "#/components/schemas/MeetingSummaryEvent",
                      "mute": "#/components/schemas/MuteEvent",
                      "spoken.answer": "#/components/schemas/SpokenAnswerEvent",
                      "stop": "#/components/schemas/StopEvent",
                      "transcript.segment": "#/components/schemas/TranscriptSegmentEvent",
                      "wake": "#/components/schemas/WakeEvent"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/BotStatusEvent"
                    },
                    {
                      "$ref": "#/components/schemas/TranscriptSegmentEvent"
                    },
                    {
                      "$ref": "#/components/schemas/WakeEvent"
                    },
                    {
                      "$ref": "#/components/schemas/StopEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MuteEvent"
                    },
                    {
                      "$ref": "#/components/schemas/AlertEvent"
                    },
                    {
                      "$ref": "#/components/schemas/SpokenAnswerEvent"
                    },
                    {
                      "$ref": "#/components/schemas/ChatPostEvent"
                    },
                    {
                      "$ref": "#/components/schemas/FollowUpEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MeetingSummaryEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MeetingEndedEvent"
                    }
                  ],
                  "title": "Response Mute Api Meetings  Meeting Id  Mute Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mute"
      }
    },
    "/api/meetings/{meeting_id}/stop": {
      "post": {
        "operationId": "stop_api_meetings__meeting_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "alert": "#/components/schemas/AlertEvent",
                      "bot.status": "#/components/schemas/BotStatusEvent",
                      "chat.post": "#/components/schemas/ChatPostEvent",
                      "follow_up": "#/components/schemas/FollowUpEvent",
                      "meeting.ended": "#/components/schemas/MeetingEndedEvent",
                      "meeting.summary": "#/components/schemas/MeetingSummaryEvent",
                      "mute": "#/components/schemas/MuteEvent",
                      "spoken.answer": "#/components/schemas/SpokenAnswerEvent",
                      "stop": "#/components/schemas/StopEvent",
                      "transcript.segment": "#/components/schemas/TranscriptSegmentEvent",
                      "wake": "#/components/schemas/WakeEvent"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/BotStatusEvent"
                    },
                    {
                      "$ref": "#/components/schemas/TranscriptSegmentEvent"
                    },
                    {
                      "$ref": "#/components/schemas/WakeEvent"
                    },
                    {
                      "$ref": "#/components/schemas/StopEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MuteEvent"
                    },
                    {
                      "$ref": "#/components/schemas/AlertEvent"
                    },
                    {
                      "$ref": "#/components/schemas/SpokenAnswerEvent"
                    },
                    {
                      "$ref": "#/components/schemas/ChatPostEvent"
                    },
                    {
                      "$ref": "#/components/schemas/FollowUpEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MeetingSummaryEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MeetingEndedEvent"
                    }
                  ],
                  "title": "Response Stop Api Meetings  Meeting Id  Stop Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop"
      }
    },
    "/api/meetings/{meeting_id}/wake": {
      "post": {
        "operationId": "wake_api_meetings__meeting_id__wake_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "alert": "#/components/schemas/AlertEvent",
                      "bot.status": "#/components/schemas/BotStatusEvent",
                      "chat.post": "#/components/schemas/ChatPostEvent",
                      "follow_up": "#/components/schemas/FollowUpEvent",
                      "meeting.ended": "#/components/schemas/MeetingEndedEvent",
                      "meeting.summary": "#/components/schemas/MeetingSummaryEvent",
                      "mute": "#/components/schemas/MuteEvent",
                      "spoken.answer": "#/components/schemas/SpokenAnswerEvent",
                      "stop": "#/components/schemas/StopEvent",
                      "transcript.segment": "#/components/schemas/TranscriptSegmentEvent",
                      "wake": "#/components/schemas/WakeEvent"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/BotStatusEvent"
                    },
                    {
                      "$ref": "#/components/schemas/TranscriptSegmentEvent"
                    },
                    {
                      "$ref": "#/components/schemas/WakeEvent"
                    },
                    {
                      "$ref": "#/components/schemas/StopEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MuteEvent"
                    },
                    {
                      "$ref": "#/components/schemas/AlertEvent"
                    },
                    {
                      "$ref": "#/components/schemas/SpokenAnswerEvent"
                    },
                    {
                      "$ref": "#/components/schemas/ChatPostEvent"
                    },
                    {
                      "$ref": "#/components/schemas/FollowUpEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MeetingSummaryEvent"
                    },
                    {
                      "$ref": "#/components/schemas/MeetingEndedEvent"
                    }
                  ],
                  "title": "Response Wake Api Meetings  Meeting Id  Wake Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wake"
      }
    },
    "/api/settings": {
      "get": {
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings"
      },
      "put": {
        "operationId": "put_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Settings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Settings"
      }
    },
    "/webhooks/recall/{token}": {
      "post": {
        "operationId": "recall_webhook_webhooks_recall__token__post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recall Webhook"
      }
    }
  }
}